Add graphql project, the first Go GraphQL client.
To my knowledge, this is the first GraphQL client package available
for Go.

It's still in active development, but by now, it's feature complete
and functional. It has support for performing queries and mutations,
including arguments, variables, aliases, inline fragments, unions, etc.

The focus of the package was to come up with a friendly, simple,
powerful API for a Go GraphQL client. The history of its development
and API design decisions can be seen at google/go-github#646.

This is a general GraphQL client. It was factored out from the
githubql project, which is a GraphQL client that targets the
GitHub GraphQL API specifically.

It has been tested via simple test programs, as well as more
advanced use cases, and comes with thorough test coverage.
Its API is becoming more stable, so I think it's a good time to
add it to the list of GraphQL-related code.

Followup to #348.